refacto types
